requirements.txt
langchain>=0.3.27,
langchain-chroma>=0.2.5,
langchain-community>=0.3.28,
langchain-openai>=0.3.32,
pypdf>=6.0.0,
python-dotenv>=1.1.1,